requirements.txt
aiohttp==3.7.3
aiohttp-jinja2==1.4.2
aiohttp-socks==0.5.5
aiovk==4.0.0
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
idna==3.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
python-environ==0.4.54
python-socks==1.2.0
typing-extensions==3.7.4.3
yarl==1.6.3